requirements.txt
streamlit==1.31.0
langchain==0.1.10
langchain-groq==0.0.1
langchainhub==0.1.15
wikipedia==1.4.0
duckduckgo-search==5.0.0
python-dotenv==1.0.1